redid the resolver to include auth

+++ voting-service/models/politicalParty/resolver.js
@@ -1,22 +1,34 @@
 const PoliticalParty = require('./schema').PoliticalParties;
+const authRoles = require('../../authRoles').resolverToRole;
 
 
-const getPoliticalParties = (parent, args) => PoliticalParty.find({});
+const getPoliticalParties = (parent, args, context) => {
+  if (!authRoles[context.payload.role].includes('getPoliticalParties')) throw new Error(`User ${context.payload.role} cannot access resolver getPoliticalParties`)
 
-const getPoliticalParty = (parent, args) => new Promise((resolve, reject) => {
+  return PoliticalParty.find({});
+
+}
+
+const getPoliticalParty = (parent, args, context) => new Promise((resolve, reject) => {
+  if (!authRoles[context.payload.role].includes('getPoliticalParty')) throw new Error(`User ${context.payload.role} cannot access resolver getPoliticalParty`)
+
   PoliticalParty.findById(args.id, (err, res) => {
     err ? reject(err) : resolve(res);
   });
 });
 
-const getPoliticalPartyByName = (parent, args) => new Promise((resolve, reject) => {
+const getPoliticalPartyByName = (parent, args, context) => new Promise((resolve, reject) => {
+  if (!authRoles[context.payload.role].includes('getPoliticalPartyByName')) throw new Error(`User ${context.payload.role} cannot access resolver getPoliticalPartyByName`)
+
   PoliticalParty.find({ name: args.name }, (err, res) => {
     if (err || res.length == 0) { return reject(res); }
     return resolve(res);
   });
 });
 
-const addPoliticalParty = (parent, args) => {
+const addPoliticalParty = (parent, args, context) => {
+  if (!authRoles[context.payload.role].includes('addPoliticalParty')) throw new Error(`User ${context.payload.role} cannot access resolver addPoliticalParty`)
+
   const newPoliticalParty = new PoliticalParty({
     name: args.name,
     colour: args.colour,
@@ -28,7 +40,9 @@ const addPoliticalParty = (parent, args) => {
   });
 };
 
-const updatePoliticalParty = (parent, args) => new Promise((resolve, reject) => {
+const updatePoliticalParty = (parent, args, context) => new Promise((resolve, reject) => {
+  if (!authRoles[context.payload.role].includes('updatePoliticalParty')) throw new Error(`User ${context.payload.role} cannot access resolver updatePoliticalParty`)
+
   PoliticalParty.findOneAndUpdate({ _id: args.id },
     { $set: { name: args.name, colour: args.colour } }, { new: true },
     (err, res) => {
@@ -36,7 +50,9 @@ const updatePoliticalParty = (parent, args) => new Promise((resolve, reject) => 
     });
 });
 
-const deletePoliticalParty = (parent, args) => new Promise((resolve, reject) => {
+const deletePoliticalParty = (parent, args, context) => new Promise((resolve, reject) => {
+  if (!authRoles[context.payload.role].includes('deletePoliticalParty')) throw new Error(`User ${context.payload.role} cannot access resolver deletePoliticalParty`)
+
   PoliticalParty.findByIdAndRemove({ _id: args.id }, (err, res) => {
     err ? reject(err) : resolve(res);
   });
